requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==24.10.0
boto3==1.35.68
botocore==1.35.68
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
Django==5.1.3
django-celery==3.1.17
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-countries==7.6.1
django-environ==0.11.2
django-filter==24.3
django-ipware==7.0.1
django-js-asset==2.2.0
django-pymemcache==1.0.0
django-storages==1.14.4
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-tracking-logger==1.9.6
drf-yasg==1.21.8
firebase-admin==6.6.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonformatter==0.3.4
kombu==5.4.2
msgpack==1.1.0
mypy-extensions==1.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
phonenumbers==8.13.50
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.0
pymemcache==4.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-firebase==1.2
python-ipware==3.0.0
python-json-logger==2.0.7
#python-magic-bin==0.4.14
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
tomli==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
